refactor(console): extract realtime orchestration owner
